update - hw5

+++ src/services/user.service.ts
@@ -1,25 +1,61 @@
-import { IUser, IUserDTO } from "../interfaces/user.interface";
+import { StatusCodesEnum } from "../enums/status-codes.enums";
+import { ApiError } from "../errors/api.errors";
+import { IUser, IUserUpdateDTO } from "../interfaces/user.interface";
 import { userRepository } from "../repositories/user.repository";
 
 class UserService {
     public getAll(): Promise<IUser[]> {
         return userRepository.getAll();
     }
-    public create(user: IUserDTO): Promise<IUser> {
-        return userRepository.create(user);
+
+    public async getById(userId: string): Promise<IUser> {
+        const user = await userRepository.getById(userId);
+        if (!user) {
+            throw new ApiError("User not fount", StatusCodesEnum.NOT_FOUND);
+        }
+        return user;
     }
-    public getById(userId: string): Promise<IUser> {
-        return userRepository.getById(userId);
-    }
-    public async updateById(id: string, updateData: IUserDTO) {
-        await userRepository.updateById(id, updateData);
+
+    public async updateById(id: string, updateData: IUserUpdateDTO) {
+        const user = await userRepository.updateById(id, updateData);
+        if (!user) {
+            throw new ApiError("User not fount", StatusCodesEnum.NOT_FOUND);
+        }
         return await userRepository.getById(id);
     }
-    public deleteById(id: string) {
-        return userRepository.deleteById(id);
+
+    public async deleteById(id: string) {
+        const user = await userRepository.getById(id);
+        if (!user) {
+            throw new ApiError("User not fount", StatusCodesEnum.NOT_FOUND);
+        }
+        await userRepository.deleteById(id);
     }
+
     public deleteAll() {
         return userRepository.deleteAll();
+    }
+
+    public async isEmailUnique(email: string): Promise<void> {
+        const user = await userRepository.getByEmail(email);
+        if (user) {
+            throw new ApiError(
+                "User is already exists",
+                StatusCodesEnum.BAD_REQUEST,
+            );
+        }
+    }
+
+    public async isActive(id: string): Promise<boolean> {
+        const user = await this.getById(id);
+        return user.isActive;
+    }
+
+    public blockUser(id: string): Promise<IUser> {
+        return userRepository.blockUser(id);
+    }
+    public unBlockUser(id: string): Promise<IUser> {
+        return userRepository.unBlockUser(id);
     }
 }
